requirements.txt
joblib==1.1.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
PyYAML==6.0
scikit_learn==1.1.1
seaborn==0.11.2
xgboost==1.6.1
